# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /api/tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/tasks/{id}:
    patch:
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      tags:
        - tasks
      requestBody:
        description: The task field to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"
        "400":
          description: Bad request for update task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"
    delete:
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      tags:
        - tasks
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"
  /api/users/{user-id}/tasks:
    get:
      summary: Get all tasks for user
      parameters:
        - name: user-id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      tags:
        - users
      responses:
        '200':
          description: User with tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{id}:
    patch:
      summary: Update user
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"
        "400":
          description: Bad request for update user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"
    delete:
      summary: Delete user
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MsgResponse"

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    MsgResponse:
      type: object
      properties:
        message:
          type: string